Extract classes to their own files
